+++ index.js
@@ -1,48 +1,11 @@
-var ejs = require('ejs');
-var fs = require('fs');
-var path = require('path');
-
-var render = ejs.render;
-
-var tagParser = require('./app/tag_parser');
-
-function getDirname(file, parentPath) {
-    var dirname = file;
-    if (parentPath) dirname = path.join(parentPath, "../", file);
-    return dirname;
-}
-
-function parseContent(dirname, options) {
-    var content = fs.readFileSync(dirname, 'utf-8');
-    if (options.ejsEscape !== false) content = content.replace(options.tagRegex, "<%%");
-    return tagParser(content, options.tokens, options);
-}
-
-function setupOptions(dirname, options) {
-    var rendererOptions = options.args || {};
-    rendererOptions.xejs = xejs;
-    rendererOptions.parentPath = dirname;
-    rendererOptions.options = options;
-    return rendererOptions;
-}
-
-function xejs(file, options, parentPath) {
-    var dirname = getDirname(file, parentPath);
-    var content = parseContent(dirname, options);
-    var rendererOptions = setupOptions(dirname, options);
-    content = render(content, rendererOptions);
-    rendererOptions.parentPath = parentPath;
-    return content;
-}
+var defaultTags = require('./app/tag_parser').defaultTags;
+var xejs = require('./app/renderer');
 
 
-module.exports = function(file, renderingOptions, args, done) {
-    var tokens = tagParser.defaultTags;
+function setOptions(renderingOptions, args) {
+    var tokens = defaultTags;
     if (renderingOptions.tokens) tokens = tokens.concat(renderingOptions.tokens);
-    if (!done && typeof args === "function") done = args;
-    else if (!done && !args && typeof renderingOptions === "function") done = renderingOptions;
-
-    var options = {
+    return {
         openTagEJS: "<%- ",
         closeTagEJS: "%>",
         tagRegex: /<%/g,
@@ -51,12 +14,27 @@ module.exports = function(file, renderingOptions, args, done) {
         commentTag: renderingOptions.commentTag || "#",
         ejsEscape: renderingOptions.ejsEscape === false ? false : true,
         tokens: tokens,
-        args: args || {}
+        args: args || {},
+        renderedStack: []
     };
+}
+
+//TODO: Avoid repeating code
+function renderFile(file, renderingOptions, args, done) {
+    if (!done && typeof args === "function") {
+        done = args;
+        args = [];
+    } else if (!done && !args && typeof renderingOptions === "function") {
+        done = renderingOptions;
+        renderingOptions = {};
+    }
+
+    var options = setOptions(renderingOptions, args);
+
     var res = null;
     var err = null;
     try {
-        res = xejs(file, options);
+        res = xejs.renderFile(file, options);
     } catch (e) {
         err = e;
     }
@@ -68,4 +46,38 @@ module.exports = function(file, renderingOptions, args, done) {
             resolve(res);
         }
     });
-};
+}
+
+// TODO: avoid repeating code
+function renderString(content, renderingOptions, args, done) {
+    if (!done && typeof args === "function") {
+        done = args;
+        args = [];
+    } else if (!done && !args && typeof renderingOptions === "function") {
+        done = renderingOptions;
+        renderingOptions = {};
+    }
+
+    var options = setOptions(renderingOptions, args);
+    var includePath=renderingOptions.includePath;
+
+    var res = null;
+    var err = null;
+    try {
+        res = xejs.renderString(content, options, includePath);
+    } catch (e) {
+        err = e;
+    }
+    if (done) return done(err, res);
+    return new Promise(function(resolve, reject) {
+        if (err) {
+            reject(err);
+        } else {
+            resolve(res);
+        }
+    });
+}
+
+module.exports = renderFile;
+module.exports.renderFile = renderFile;
+module.exports.renderString = renderString;