+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.6.0 / 2016-10-22
+==================
+
+  * Error thrown if circular dependencies on include are detected
+  * Support for rendering from string
+  * Node.js updated in CI tests
+
 0.5.0 / 2016-10-15
 ==================